Move the local storage checks into a component
