+++ layouts/default.vue
@@ -252,13 +252,7 @@
 
 
     <nuxt ref="pageContent" class="ml-0 pl-0 pl-sm-1 pr-0 pr-sm-1 pageContent" />
-    <div v-if="localStorageErrors" class="storage w-100 text-center">
-      <b-alert show variant="danger">
-        <p>We can't access local storage on your browser.</p>
-        <p>Please clear your cache for this site. If you have security software, please disable it for this site.</p>
-        <p>We'll carry on, but things may go wrong...</p>
-      </b-alert>
-    </div>
+    <LocalStorageMonitor />
     <client-only>
       <ChatPopups v-if="loggedIn" class="d-none d-sm-block" />
       <LoginModal ref="loginModal" />
@@ -280,6 +274,7 @@
 <script>
 // Import login modal as I've seen an issue where it's not in $refs when you click on the signin button too rapidly.
 import LoginModal from '~/components/LoginModal'
+import LocalStorageMonitor from '~/components/LocalStorageMonitor'
 
 const AboutMeModal = () => import('~/components/AboutMeModal')
 const ChatPopups = () => import('~/components/ChatPopups')
@@ -293,7 +288,8 @@ export default {
     ChatPopups,
     Notification,
     AboutMeModal,
-    LoginModal
+    LoginModal,
+    LocalStorageMonitor
   },
 
   data: function() {
@@ -302,8 +298,7 @@ export default {
       distance: 1000,
       notificationPoll: null,
       nchan: null,
-      logo: require(`@/static/icon.png`),
-      localStorageErrors: false
+      logo: require(`@/static/icon.png`)
     }
   },
 
@@ -446,8 +441,6 @@ export default {
     } catch (e) {
       console.log('Failed to set context', e)
     }
-
-    this.monitorLocalStorage()
   },
 
   async beforeCreate() {
@@ -652,57 +645,6 @@ export default {
       await this.$store.dispatch('notifications/allSeen')
       await this.$store.dispatch('notifications/count')
       await this.$store.dispatch('notifications/list')
-    },
-
-    monitorLocalStorage() {
-      // We have trouble on some devices setting info to localStorage, due to quota or security reasons.  This can
-      // break us.  Try to set something to local storage so that we can check if it makes it there;
-      // if not, then do a toast.
-      console.log('Monitor local storage')
-      const now = Date.now()
-      this.$store.dispatch('misc/set', {
-        key: 'localStorageMonitor',
-        value: now
-      })
-
-      setTimeout(() => {
-        // Go directly to local storage to see if it's made it.
-        console.log('Now check if it made it')
-        let ok = false
-
-        try {
-          console.log('Get data')
-          const stored = localStorage.getItem('iznik')
-          console.log('Parse stored')
-          const decoded = JSON.parse(stored)
-          console.log('Decoded')
-
-          if (decoded && decoded.misc && decoded.misc.localStorageMonitor) {
-            console.log('Got back', decoded.misc.localStorageMonitor)
-            const age =
-              now - new Date(decoded.misc.localStorageMonitor).getTime()
-            console.log('Age', age)
-
-            if (age < 60000) {
-              console.log('Ok')
-              ok = true
-            }
-          } else {
-            console.log("Doesn't have what we expect", decoded)
-          }
-        } catch (e) {
-          console.log('Failed to parse local storage')
-        }
-
-        if (!ok) {
-          console.error('Errors with local storage')
-          this.localStorageErrors = true
-
-          setTimeout(() => {
-            this.localStorageErrors = false
-          }, 30000)
-        }
-      }, 30000)
     }
   }
 }
@@ -909,11 +851,6 @@ svg.fa-icon {
     visibility: visible;
     opacity: 1;
   }
-}
-
-.storage {
-  position: fixed;
-  bottom: 0%;
 }
 
 .toggler {
